% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{TCGA_BLCA}
\alias{TCGA_BLCA}
\title{Read counts and dispersion distribution for genes in TCGA BLCA data.}
\usage{
TCGA_BLCA
}
\value{
A class 'DGEList' from edgeR package.
}
\description{
Read counts and dispersion distribution for genes in TCGA BLCA data.
}

