% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{TCGA_BRCA}
\alias{TCGA_BRCA}
\title{Read counts and dispersion distribution for genes in TCGA BRCA data.}
\source{
\url{https://tcga-data.nci.nih.gov/tcgafiles/ftp_auth/distro_ftpusers/anonymous/tumor/brca/cgcc/unc.edu/illuminahiseq_rnaseqv2/rnaseqv2/unc.edu_BRCA.IlluminaHiSeq_RNASeqV2.Level_3.1.10.0/}
}
\usage{
TCGA_BRCA
}
\value{
A class 'DGEList' from edgeR package.
}
\description{
Read counts and dispersion distribution for genes in TCGA BRCA data.
}

