% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{TCGA_LGG}
\alias{TCGA_LGG}
\title{Read counts and dispersion distribution for genes in TCGA LGG data.}
\usage{
TCGA_LGG
}
\value{
A class 'DGEList' from edgeR package.
}
\description{
Read counts and dispersion distribution for genes in TCGA LGG data.
}

