% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{TCGA_LUAD}
\alias{TCGA_LUAD}
\title{Read counts and dispersion distribution for genes in TCGA LUAD data.}
\source{
\url{https://tcga-data.nci.nih.gov/tcgafiles/ftp_auth/distro_ftpusers/anonymous/tumor/luad/cgcc/unc.edu/illuminahiseq_rnaseqv2/rnaseqv2/unc.edu_LUAD.IlluminaHiSeq_RNASeqV2.Level_3.1.13.0/}
}
\usage{
TCGA_LUAD
}
\value{
A class 'DGEList' from edgeR package.
}
\description{
Read counts and dispersion distribution for genes in TCGA LUAD data.
}

