% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{TCGA_BLCA}
\alias{TCGA_BLCA}
\title{Read counts and dispersion distribution for genes in TCGA BLCA data.}
\source{
\url{https://tcga-data.nci.nih.gov/tcgafiles/ftp_auth/distro_ftpusers/anonymous/tumor/blca/cgcc/unc.edu/illuminahiseq_rnaseqv2/rnaseqv2/unc.edu_BLCA.IlluminaHiSeq_RNASeqV2.Level_3.1.17.0/}
}
\usage{
TCGA_BLCA
}
\value{
A class 'DGEList' from edgeR package.
}
\description{
Read counts and dispersion distribution for genes in TCGA BLCA data.
}

